two more hill climbing
