+++ src/hyperactive/opt/__init__.py
@@ -1,10 +1,11 @@
 """Individual optimization algorithms."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
 
-from hyperactive.opt.gfo import HillClimbing
+from hyperactive.opt.hillclimbing import HillClimbing, StochasticHillClimbing
 from hyperactive.opt.gridsearch import GridSearch
 
 __all__ = [
     "GridSearch",
     "HillClimbing",
+    "StochasticHillClimbing",
 ]
