+++ src/hyperactive/opt/_adapters/__init__.py
@@ -1,0 +1,2 @@
+"""Adapters for individual packages."""
+# copyright: hyperactive developers, MIT License (see LICENSE file)
