+++ src/hyperactive/opt/_adapters/_gfo.py
@@ -1,0 +1,116 @@
+"""Adapter for gfo package."""
+# copyright: hyperactive developers, MIT License (see LICENSE file)
+
+from gradient_free_optimizers import HillClimbingOptimizer
+from hyperactive.base import BaseOptimizer
+from skbase.utils.stdout_mute import StdoutMute
+
+
+class _BaseGFOadapter(BaseOptimizer):
+    """Adapter base class for gradient-free-optimizers.
+
+    * default tag setting
+    * Handles defaults for "initialize"
+    * provides default get_search_config
+    * provides default get_test_params
+    * extension interface: _get_gfo_class
+    """
+
+    _tags = {
+        "authors": "SimonBlanke",
+        "python_dependencies": ["gradient-free-optimizers>=1.5.0"],
+    }
+
+    def __init__(self):
+
+        super().__init__()
+
+        if self.initialize is None:
+            self._initialize = {"grid": 4, "random": 2, "vertices": 4}
+        else:
+            self._initialize = self.initialize
+
+    def _get_gfo_class(self):
+        """Get the GFO class to use.
+
+        Returns
+        -------
+        class
+            The GFO class to use. One of the concrete GFO classes
+        """
+        raise NotImplementedError(
+            "This method should be implemented in a subclass."
+        )
+
+    def get_search_config(self):
+        """Get the search configuration.
+
+        Returns
+        -------
+        dict with str keys
+            The search configuration dictionary.
+        """
+        search_config = super().get_search_config()
+        search_config["initialize"] = self._initialize
+        del search_config["verbose"]
+        return search_config
+
+    @classmethod
+    def get_test_params(cls, parameter_set="default"):
+        """Return testing parameter settings for the skbase object.
+
+        ``get_test_params`` is a unified interface point to store
+        parameter settings for testing purposes. This function is also
+        used in ``create_test_instance`` and ``create_test_instances_and_names``
+        to construct test instances.
+
+        ``get_test_params`` should return a single ``dict``, or a ``list`` of ``dict``.
+
+        Each ``dict`` is a parameter configuration for testing,
+        and can be used to construct an "interesting" test instance.
+        A call to ``cls(**params)`` should
+        be valid for all dictionaries ``params`` in the return of ``get_test_params``.
+
+        The ``get_test_params`` need not return fixed lists of dictionaries,
+        it can also return dynamic or stochastic parameter settings.
+
+        Parameters
+        ----------
+        parameter_set : str, default="default"
+            Name of the set of test parameters to return, for use in tests. If no
+            special parameters are defined for a value, will return `"default"` set.
+
+        Returns
+        -------
+        params : dict or list of dict, default = {}
+            Parameters to create testing instances of the class
+            Each dict are parameters to construct an "interesting" test instance, i.e.,
+            `MyClass(**params)` or `MyClass(**params[i])` creates a valid test instance.
+            `create_test_instance` uses the first (or only) dictionary in `params`
+        """
+        import numpy as np
+        from hyperactive.experiment.integrations import SklearnCvExperiment
+
+        sklearn_exp = SklearnCvExperiment.create_test_instance()
+        params_sklearn = {
+            "experiment": sklearn_exp,
+            "search_space": {
+                "C": np.array([0.01, 0.1, 1, 10]),
+                "gamma": np.array([0.0001, 0.01, 0.1, 1, 10]),
+            },
+            "n_iter": 100,
+        }
+
+        from hyperactive.experiment.toy import Ackley
+
+        ackley_exp = Ackley.create_test_instance()
+        params_ackley = {
+            "experiment": ackley_exp,
+            "search_space": {
+                "x0": np.linspace(-5, 5, 10),
+                "x1": np.linspace(-5, 5, 10),
+            },
+            "n_iter": 100,
+        }
+        
+        return [params_sklearn, params_ackley]
