--- src/hyperactive/opt/gfo/__init__.py
@@ -1,6 +1,0 @@
-"""Optimizers from Gradient free optimizers package."""
-# copyright: hyperactive developers, MIT License (see LICENSE file)
-
-from hyperactive.opt.gfo.hillclimbing import HillClimbing
-
-__all__ = ["HillClimbing"]
