+++ src/hyperactive/opt/gridsearch.py
@@ -22,7 +22,7 @@ class GridSearch(BaseOptimizer):
         The score to assign if an error occurs during the evaluation of a parameter set.
     experiment : BaseExperiment, optional
         The experiment to optimize parameters for.
-        Optional, can be passed later in ``add_search``.
+        Optional, can be passed later via ``set_params``.
 
     Example
     -------
