+++ src/hyperactive/opt/hillclimbing/__init__.py
@@ -1,0 +1,10 @@
+"""Optimizers from Gradient free optimizers package."""
+# copyright: hyperactive developers, MIT License (see LICENSE file)
+
+from hyperactive.opt.hillclimbing._hillclimbing import HillClimbing
+from hyperactive.opt.hillclimbing._stochastic import StochasticHillClimbing
+
+__all__ = [
+    "HillClimbing",
+    "StochasticHillClimbing",
+]
